requirements.txt
Django
Pillow
django-celery-beat
celery
redis
django-haystack
Whoosh
django-paypal
psycopg2-binary
gunicorn
dj-database-url
python-dotenv
whitenoise
watchgod
pytest
pytest-django
pytest-xdist
factory-boy
flake8
